Add UnresolvedInstanceException
