+++ src/main/java/jp/mydns/projectk/safi/UnresolvedInstanceException.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2025, Project-K
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * * Redistributions of source code must retain the above copyright notice, this
+ *   list of conditions and the following disclaimer.
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+package jp.mydns.projectk.safi;
+
+import java.util.Objects;
+
+/**
+ Exception where a CDI bean instance cannot be resolved. Raised only when a dependent instance
+ cannot be resolved, this exception indicates an illegal state caused by an implementation bug. The
+ caller should not proceed further.
+
+ @author riru
+ @version 3.0.0
+ @since 3.0.0
+ */
+public class UnresolvedInstanceException extends PublishableRuntimeException {
+
+@java.io.Serial
+private static final long serialVersionUID = 7274928374928374907L;
+
+/**
+ Construct with the {@code Throwable}.
+
+ @param cause the {@code Throwable} for maintainer.
+ @throws NullPointerException if {@code cause} is {@code null}
+ @since 3.0.0
+ */
+public UnresolvedInstanceException(Throwable cause) {
+    super(Objects.requireNonNull(cause));
+}
+
+}
